requirements.txt
streamlit
gpxpy
qrcode[pil]
Pillow
